Add additional bridging for use by SkipFuseUI

+++ Sources/SkipUI/SkipUI/Containers/VStack.swift
@@ -36,9 +36,9 @@ public struct VStack : View {
     }
 
     // SKIP @bridge
-    public init(alignmentKey: String, spacing: Double?, bridgedContent: any View) {
+    public init(alignmentKey: String, spacing: CGFloat?, bridgedContent: any View) {
         self.alignment = HorizontalAlignment(key: alignmentKey)
-        self.spacing = spacing == nil ? nil : CGFloat(spacing!)
+        self.spacing = spacing == nil ? nil : spacing!
         self.content = ComposeBuilder.from { bridgedContent }
     }
